Fix detail page review button logic (again)
can_rate and has_rated are available, developed is not.

This fix shows the appropriate buttons in any circumstance where it's possible for the user to have access to use either of the buttons. When the user logs in, if there is an issue which would otherwise prevent them from being able to use the buttons, the page will refresh anyway and the issue will immediately correct itself.