Update to jQuery 2.0.2, remove JQ from the builder (bug 877906)
Note that I left in the following modules:
- css
- offset

We're only using scrollTop in one place and offset in one other. Once
those are ported, we can remove the offset module. We use the css()
method in only a few places, and that could be ported easily as well,
removing that dependency. The offset module depends on the css module.