Added title attrs to facilitate reading long lines (bug 1079172).
I’ve added title attributes to anything with an elliflow class, and
also some img elements.

On the market_tile macro there is a problem where the .info div
displays above the icon, so the tooltip doesn’t display on mouseover.
The .icon and .info classes sound very generic, so I didn’t want to
change their styles to fix this problem because I fear causing further
problems. Someone more familiar with the styles may want to point me
in the right direction.